refactor: animationコンポーネントの配置を修正
