+++ src/animations/index.ts
@@ -1,0 +1,6 @@
+import { CreateDriftingAnimation } from "./CreateDriftingAnimation"
+import { CreateJumpAnimation } from "./CreateJumpAnimation"
+import { CreateLoadingAnimation } from "./CreateLoadingAnimation"
+import { CreateMovingWindowFromRightAnimation } from "./CreateMovingWindowFromRightAnimation"
+import { CreateShakeAndUpAnimation } from "./CreateShakeAndUpAnimation"
+export { CreateDriftingAnimation, CreateJumpAnimation, CreateLoadingAnimation, CreateMovingWindowFromRightAnimation, CreateShakeAndUpAnimation }
